- fix up assets tree shaking - rename `group` to `isUniformGroup` to avoid clash with new iOS property
